requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
requests==2.31.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-ratelimit==4.1.0
gunicorn==21.2.0
whitenoise==6.5.0

# Testing (optional)
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
